+++ flutter_ffi_plugin/example/native/sample_crate/src/lib.rs
@@ -1,39 +1,42 @@
 //! This crate is written for Rinf demonstrations.
 
-pub use fractal::draw_fractal_image;
-
+mod error;
 mod fractal;
 
-// `machineid_rs` only supports desktop platforms.
+use error::ExampleError;
 
+pub use fractal::draw_fractal_image;
+
+// `machineid_rs` only supports desktop platforms.
 #[cfg(any(target_os = "windows", target_os = "macos", target_os = "linux"))]
-pub fn get_hardward_id() -> Option<String> {
+pub fn get_hardward_id() -> Result<String, ExampleError> {
     let mut builder = machineid_rs::IdBuilder::new(machineid_rs::Encryption::MD5);
     builder
         .add_component(machineid_rs::HWIDComponent::SystemID)
         .add_component(machineid_rs::HWIDComponent::CPUCores);
-    let hwid = builder.build("mykey").unwrap();
-    Some(hwid)
+    let hwid = builder
+        .build("mykey")
+        .map_err(|error| ExampleError(error.into()))?;
+    Ok(hwid)
 }
 #[cfg(not(any(target_os = "windows", target_os = "macos", target_os = "linux")))]
-pub fn get_hardward_id() -> Option<String> {
-    None
+pub fn get_hardward_id() -> Result<String, ExampleError> {
+    Ok(String::from("UNSUPPORTED"))
 }
 
 // `chrono` supports all platforms, including web.
-
 use chrono::{offset, DateTime};
 pub fn get_current_time() -> DateTime<offset::Local> {
     offset::Local::now()
 }
 
 // `reqwest` supports all platforms, including web.
-
-pub async fn fetch_from_web_api(url: &str) -> String {
-    reqwest::get(url)
+pub async fn fetch_from_web_api(url: &str) -> Result<String, ExampleError> {
+    let fetched = reqwest::get(url)
         .await
-        .expect("Could not get the response from the example web API.")
+        .map_err(|error| ExampleError(error.into()))?
         .text()
         .await
-        .expect("Could not read body from the web response.")
+        .map_err(|error| ExampleError(error.into()))?;
+    Ok(fetched)
 }