+++ flutter_ffi_plugin/lib/rinf.dart
@@ -13,28 +13,18 @@ export 'src/interface.dart' show RustSignal;
 /// This function might not be necessary for major platforms
 /// but can be useful when the app runs on embedded devices.
 void setCompiledLibPath(String? path) {
-  setCompiledLibPathExtern(path);
+  setCompiledLibPathReal(path);
 }
 
 /// Prepares the native interface
 /// needed to communicate with Rust.
 Future<void> prepareInterface(HandleRustSignal handleRustSignal) async {
-  await prepareInterfaceExtern(handleRustSignal);
+  await prepareInterfaceReal(handleRustSignal);
 }
 
 /// Starts the `main` function in Rust.
 void startRustLogic() async {
-  startRustLogicExtern();
-}
-
-/// Terminates all Rust tasks.
-/// Calling this function before closing the Flutter app
-/// can prevent potential resource leaks that may occur
-/// if the Rust side is abruptly terminated.
-/// Please note that on the web, this function does not have any effect,
-/// as tasks are managed by the JavaScript runtime, not Rust.
-void stopRustLogic() async {
-  stopRustLogicExtern();
+  startRustLogicReal();
 }
 
 /// Sends a signal to Rust.
@@ -43,7 +33,7 @@ void sendDartSignal(
   Uint8List messageBytes,
   Uint8List binary,
 ) async {
-  sendDartSignalExtern(
+  sendDartSignalReal(
     messageId,
     messageBytes,
     binary,