+++ rust_crate/src/interface_os.rs
@@ -1,29 +1,131 @@
-use super::SharedCell;
-use allo_isolate::IntoDart;
-use allo_isolate::Isolate;
-use allo_isolate::ZeroCopyBuffer;
+use crate::error::RinfError;
+use allo_isolate::{IntoDart, Isolate, ZeroCopyBuffer};
+use os_thread_local::ThreadLocal;
 use std::cell::RefCell;
-use std::panic::catch_unwind;
-use std::sync::Mutex;
-use std::sync::OnceLock;
+use std::future::Future;
+use std::pin::Pin;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::{Arc, Mutex, OnceLock};
+use std::task::{Context, Poll, Waker};
+use std::thread;
+use tokio::runtime::Builder;
 
-static DART_ISOLATE: SharedCell<Isolate> = OnceLock::new();
+static DART_ISOLATE: Mutex<Option<Isolate>> = Mutex::new(None);
 
 #[no_mangle]
 pub extern "C" fn prepare_isolate_extern(port: i64) {
-    let _ = catch_unwind(|| {
-        let dart_isolate = Isolate::new(port);
-        let cell = DART_ISOLATE
-            .get_or_init(|| Mutex::new(RefCell::new(None)))
-            .lock()
-            .unwrap();
-        cell.replace(Some(dart_isolate));
-    });
+    let dart_isolate = Isolate::new(port);
+    let mut guard = match DART_ISOLATE.lock() {
+        Ok(inner) => inner,
+        Err(_) => {
+            let error = RinfError::LockDartIsolate;
+            println!("{error}");
+            return;
+        }
+    };
+    guard.replace(dart_isolate);
 }
 
-pub fn send_rust_signal_extern(message_id: i32, message_bytes: Vec<u8>, binary: Vec<u8>) {
-    let cell = DART_ISOLATE.get().unwrap().lock().unwrap();
-    let dart_isolate = cell.borrow().unwrap();
+// We use `os_thread_local` so that when the program fails
+// and the main thread exits unexpectedly,
+// the whole async tokio runtime can shut down as well
+// by receiving a signal via the shutdown channel.
+// Without this solution, zombie threads inside the tokio runtime
+// might outlive the app.
+type ShutdownSenderLock = OnceLock<ThreadLocal<RefCell<Option<ShutdownSender>>>>;
+static SHUTDOWN_SENDER: ShutdownSenderLock = OnceLock::new();
+
+pub fn start_rust_logic_real<F>(main_future: F) -> Result<(), RinfError>
+where
+    F: Future<Output = ()> + Send + 'static,
+{
+    // Enable backtrace output for panics.
+    #[cfg(debug_assertions)]
+    {
+        #[cfg(not(feature = "backtrace"))]
+        {
+            std::panic::set_hook(Box::new(|panic_info| {
+                crate::debug_print!("A panic occurred in Rust.\n{panic_info}");
+            }));
+        }
+        #[cfg(feature = "backtrace")]
+        {
+            std::panic::set_hook(Box::new(|panic_info| {
+                let backtrace = backtrace::Backtrace::new();
+                crate::debug_print!("A panic occurred in Rust.\n{panic_info}\n{backtrace:?}");
+            }));
+        }
+    }
+
+    // Prepare the channel that will notify tokio runtime to shutdown
+    // after the main Dart thread has gone.
+    let (shutdown_sender, shutdown_receiver) = shutdown_channel();
+    let shutdown_sender_lock =
+        SHUTDOWN_SENDER.get_or_init(move || ThreadLocal::new(|| RefCell::new(None)));
+    shutdown_sender_lock.with(|cell| cell.replace(Some(shutdown_sender)));
+
+    // Build the tokio runtime.
+    #[cfg(not(feature = "multi-worker"))]
+    {
+        let tokio_runtime = Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .map_err(|_| RinfError::BuildRuntime)?;
+        thread::spawn(move || {
+            tokio_runtime.spawn(main_future);
+            tokio_runtime.block_on(shutdown_receiver);
+            // Dropping the tokio runtime makes it shut down.
+            drop(tokio_runtime);
+        });
+    }
+    #[cfg(feature = "multi-worker")]
+    {
+        static TOKIO_RUNTIME: Mutex<Option<tokio::runtime::Runtime>> = Mutex::new(None);
+        let tokio_runtime = Builder::new_multi_thread()
+            .enable_all()
+            .build()
+            .map_err(|_| RinfError::BuildRuntime)?;
+        tokio_runtime.spawn(async {
+            main_future.await;
+        });
+        tokio_runtime.spawn(async {
+            shutdown_receiver.await;
+            thread::spawn(|| {
+                if let Ok(mut guard) = TOKIO_RUNTIME.lock() {
+                    let runtime_option = guard.take();
+                    if let Some(runtime) = runtime_option {
+                        // Dropping the tokio runtime makes it shut down.
+                        drop(runtime);
+                    }
+                }
+            })
+        });
+        if let Ok(mut guard) = TOKIO_RUNTIME.lock() {
+            // If there was already a tokio runtime previously,
+            // most likely due to Dart's hot restart,
+            // its tasks as well as itself will be terminated,
+            // being replaced with the new one.
+            let runtime_option = guard.replace(tokio_runtime);
+            if let Some(previous_runtime) = runtime_option {
+                drop(previous_runtime);
+            }
+        }
+    }
+
+    Ok(())
+}
+
+pub fn send_rust_signal_real(
+    message_id: i32,
+    message_bytes: Vec<u8>,
+    binary: Vec<u8>,
+) -> Result<(), RinfError> {
+    // When `DART_ISOLATE` is not initialized, just return the error.
+    // This can happen when running test code in Rust.
+    let guard = DART_ISOLATE
+        .lock()
+        .map_err(|_| RinfError::LockDartIsolate)?;
+    let dart_isolate = guard.as_ref().ok_or(RinfError::NoDartIsolate)?;
 
     // If a `Vec<u8>` is empty, we can't just simply send it to Dart
     // because panic can occur from null pointers.
@@ -47,4 +149,54 @@ pub fn send_rust_signal_extern(message_id: i32, message_bytes: Vec<u8>, binary: 
         ]
         .into_dart(),
     );
+
+    Ok(())
+}
+
+struct ShutdownSender {
+    is_sent: Arc<AtomicBool>,
+    waker: Arc<Mutex<Option<Waker>>>,
+}
+
+impl Drop for ShutdownSender {
+    fn drop(&mut self) {
+        self.is_sent.store(true, Ordering::SeqCst);
+        if let Ok(mut guard) = self.waker.lock() {
+            if let Some(waker) = guard.take() {
+                waker.wake();
+            }
+        }
+    }
+}
+
+struct ShutdownReceiver {
+    is_sent: Arc<AtomicBool>,
+    waker: Arc<Mutex<Option<Waker>>>,
+}
+
+impl Future for ShutdownReceiver {
+    type Output = ();
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        if self.is_sent.load(Ordering::SeqCst) {
+            Poll::Ready(())
+        } else {
+            if let Ok(mut guard) = self.waker.lock() {
+                guard.replace(cx.waker().clone());
+            }
+            Poll::Pending
+        }
+    }
+}
+
+fn shutdown_channel() -> (ShutdownSender, ShutdownReceiver) {
+    let is_sent = Arc::new(AtomicBool::new(false));
+    let waker = Arc::new(Mutex::new(None));
+
+    let sender = ShutdownSender {
+        is_sent: Arc::clone(&is_sent),
+        waker: Arc::clone(&waker),
+    };
+    let receiver = ShutdownReceiver { is_sent, waker };
+
+    (sender, receiver)
 }