+++ rust_crate/src/lib.rs
@@ -1,6 +1,4 @@
-pub use interface::*;
-
-pub mod externs;
+mod error;
 mod macros;
 
 mod interface;
@@ -8,3 +6,6 @@ mod interface;
 mod interface_os;
 #[cfg(target_family = "wasm")]
 mod interface_web;
+
+pub use error::*;
+pub use interface::*;